Explosion - Second SpinSystem Fix Attempt

+++ Project/Motor2D/M_Explosion.h
@@ -32,7 +32,7 @@ public:
 
 	iPoint position;
 	int radius;
-	int innerRadius;
+	float innerRadius;
 	int damage;
 	int nTicks;
 	float tickDelay;
@@ -51,7 +51,7 @@ struct StoredExplosion
 
 	iPoint position;
 	int radius;
-	int innerRadius;
+	float innerRadius;
 	int damage;
 	int nTicks;
 	float tickDelay;
@@ -66,7 +66,7 @@ class ExplosionSystem
 public:
 	ExplosionSystem();
 	ExplosionSystem(Unit_Type _toSpawn);
-	void PushExplosion(float delay, iPoint relativePos, int radius, int damage, int nTicks = 1, float tickDelay = 4.0f, Player_Type objective = PLAYER, bool showStencil = true, e_Explosion_Types graphic = EXPLOSION_DEFAULT, int innerRadius = 0);
+	void PushExplosion(float delay, iPoint relativePos, int radius, int damage, int nTicks = 1, float tickDelay = 4.0f, Player_Type objective = PLAYER, bool showStencil = true, e_Explosion_Types graphic = EXPLOSION_DEFAULT, float innerRadius = 0.0f);
 
 	bool Update(float dt);
 	void SetSpawningUnit(Unit_Type _toSpawn);
@@ -92,7 +92,7 @@ public:
 
 	bool CleanUp();
 
-	void AddExplosion(iPoint position, int radius, int damage, float delay = 4.0f, int nTicks = 1, Player_Type objective = PLAYER, e_Explosion_Types graphic = EXPLOSION_DEFAULT, bool showStencil = true, int innerRadius = 0);
+	void AddExplosion(iPoint position, int radius, int damage, float delay = 4.0f, int nTicks = 1, Player_Type objective = PLAYER, e_Explosion_Types graphic = EXPLOSION_DEFAULT, bool showStencil = true, float innerRadius = 0.0f);
 
 	void AddSystem(ExplosionSystem toPush, iPoint pos);
